% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_connection}
\alias{check_connection}
\title{Check if connection to server is possible}
\usage{
check_connection()
}
\value{
TRUE if connection active, else FALSE
}
\description{
Check if connection to server is possible
}
