% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_active_workers}
\alias{get_active_workers}
\title{Get the container ids of all current workers}
\usage{
get_active_workers()
}
\value{
character vector containing the unique ids of the container processes.
}
\description{
Get the container ids of all current workers
}
