% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_host}
\alias{set_host}
\title{Register the address of the host server}
\usage{
set_host(url)
}
\arguments{
\item{url}{url of the master host}
}
\value{
Exits silently
}
\description{
Register the address of the host server
}
