% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_usr_pwd}
\alias{set_usr_pwd}
\title{Set the username/password to access the API}
\usage{
set_usr_pwd(password, username)
}
\arguments{
\item{password}{password used to authenticate with the API}

\item{username}{username used to authenticate with the API}
}
\value{
Exits silently
}
\description{
Set the username/password to access the API
}
