% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{simulations_completed_info}
\alias{simulations_completed_info}
\title{Fetch information about the number of allocated containers, number of completed etc.}
\usage{
simulations_completed_info()
}
\value{
list containing three elements
\describe{
    \item{allocated}{number of containers that are currently allocated but not yet finished}
    \item{finished}{number of containers that are completely finished}
    \item{finished_past_day}{number of containers that have finished in the past 24 hours}
}
}
\description{
Fetch information about the number of allocated containers, number of completed etc.
}
