% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{register_simulation_outcomes}
\alias{register_simulation_outcomes}
\title{Send the results of the simulation to the master node}
\usage{
register_simulation_outcomes(
  scenario_uid,
  iteration_uid,
  emiss_mu_bar,
  gamma_int_bar,
  emiss_var_bar,
  emiss_varmu_bar,
  credible_interval,
  uid = getOption("sleepsimR_uid")
)
}
\arguments{
\item{scenario_uid}{string. unique id of the current simulation scenario}

\item{iteration_uid}{string. unique id of the iteration of the current scenario}

\item{emiss_mu_bar}{list. The length of this list is equal to the number of dependent variables. Each element of the list is a numeric vector that is equal to the number of hidden states, the value of which is the Maximum A Posteriori (MAP) estimate of that parameter.}

\item{gamma_int_bar}{numeric vector. m x m values where m is the number of hidden states.}

\item{emiss_var_bar}{list. The length of this list is equal to the number of dependent variables. Each element of the list is a numeric vector that is equal to the number of hidden states, the value of which is the Maximum A Posteriori (MAP) estimate of that parameter.}

\item{emiss_varmu_bar}{list. The length of this list is equal to the number of dependent variables. Each element of the list is a numeric vector that is equal to the number of hidden states, the value of which is the Maximum A Posteriori (MAP) estimate of that parameter.}

\item{credible_interval}{list. The length of this list is equal to the number of dependent variables plus one. The elements of the list are (in this order): (1) a list containing m x m elements with the lower and upper 95\% CI of the between-subject TPM intercepts (gamma_int_bar), (2) n_dep lists containing m elements with the lower and upper 95\% CI of the between-subject emission distributions.}

\item{uid}{unique id of this container}
}
\value{
if successful, a message telling the container to shut down.
}
\description{
Send the results of the simulation to the master node
}
